requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.9.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fasteners==0.14.1
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
future==0.18.2
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
pillow==11.1.0
platformdirs==4.3.6
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pysolr==3.9.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-xlib==0.29
pytz==2025.1
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
supervisor==4.2.1
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
usb-creator==0.3.7
virtualenv==20.29.2
wadllib==1.3.6
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
